feat(#582): refactor clone method to specific dashboard service
